+++ R/process_all_lyrics.R
@@ -1,21 +1,22 @@
 #' process_all_lyrics
 #'
+#' @param folder_lyrics_artists Location of json files for full artists
+#' @param folder_lyrics_songs Location of json files for songs
 #' @param daemons_artists cores to process artists json lyrics
 #' @param daemons_songs cores to process songs json lyrics
 #' @param write_output TRUE/FALSE
+#' @param filename_output Name of output file
 #' @param language Filter by language? Most common: en, es, fr, pt, ko
 #'
 #' @returns A DF with all artits and individual songs
 #' @export
 #'
 #' @examples process_all_lyrics(write_output = TRUE)
-process_all_lyrics <- function(daemons_artists = 10, daemons_songs = 5, write_output = FALSE, filename_output = NULL, language = NULL) {
-
-  # daemons_artists = 10
-  # daemons_songs = 5
-  # write_output = FALSE
-  # filename_output = NULL
-  # language = NULL
+process_all_lyrics <- function(folder_lyrics_artists = "outputs/lyrics/",
+                               folder_lyrics_songs = "outputs/lyrics_individual_songs/",
+                               daemons_artists = 10, daemons_songs = 5,
+                               write_output = FALSE, filename_output = NULL,
+                               language = NULL) {
 
   if (is.null(filename_output)) {
     filename_output = "outputs/DF_lyrics/DF_lyrics_ALL.gz"
@@ -25,8 +26,8 @@ process_all_lyrics <- function(daemons_artists = 10, daemons_songs = 5, write_ou
   }
 
   # Location of json files
-  lyrics_artists = list.files("outputs/lyrics/", pattern = "json", full.names = TRUE)
-  lyrics_songs   = list.files("outputs/lyrics_individual_songs/", pattern = "json", full.names = TRUE)
+  lyrics_artists = list.files(folder_lyrics_artists, pattern = "json", full.names = TRUE)
+  lyrics_songs   = list.files(folder_lyrics_songs, pattern = "json", full.names = TRUE)
 
 
   cli::cli_alert_info("Processing {length(lyrics_artists)} artists and {length(lyrics_songs)} songs. This  will take a minute.")
@@ -62,6 +63,7 @@ process_all_lyrics <- function(daemons_artists = 10, daemons_songs = 5, write_ou
 
   # Write
   if (write_output) {
+    if (!dir.exists(dirname(filename_output))) dir.create(dirname(filename_output))
     data.table::fwrite(DF, file = filename_output, nThread = daemons_songs)
   }