+++ run.R
@@ -6,17 +6,16 @@
 devtools::load_all()
 
 
-
 # GET Top50 songs ---------------------------------------------------------
 
-  Top50 = get_and_process(WEB = "https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9")
-  download_all_songs(Top50)
+  Top50 = get_and_process_spotify_list(spotify_list_URL = "https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9")
+  download_individual_songs(processed_spotify_list = Top50)
 
 
 # Process all lyrics ------------------------------------------------------
 
   DF_ALL_es = process_all_lyrics(write_output = TRUE, language = "es")
-  DF_ALL_es = process_all_lyrics(write_output = TRUE)
+  DF_ALL = process_all_lyrics(write_output = TRUE)
 
 
 
@@ -27,18 +26,23 @@ devtools::load_all()
   filename = here::here("outputs/DF_lyrics/DF_lyrics_ALL_es.gz")
   DF_ALL_es = data.table::fread(filename) |> tibble::as_tibble()
 
+
   set.seed(17)
   DF_ALL_es |>
     dplyr::slice_sample(n = n_subset) |>
     data.table::fwrite("outputs/DF_lyrics/DF_lyrics_es.gz")
 
+  DF_paragraphs = separate_paragraphs(DF_ALL_es)
 
+  save_experiment_materials(DF_paragraphs = DF_paragraphs,
+                            sample_n = 10,
+                            filename_output = "outputs/stimuli.js")
 
 # Get ALL songs from ALL artists ----------------------------------------------
 
   # Use this function with caution
   # Top50 Spain
-  download_process_spotify_list("https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9",
+  download_process_spotify_list(spotify_list_URL = "https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9",
                                 only_new = FALSE,
                                 continue_after_error = FALSE,
                                 end_message = TRUE)
@@ -61,20 +65,20 @@ devtools::load_all()
 
   # 5) Process
     lyrics = list.files("outputs/lyrics/", pattern = "json", full.names = TRUE)
-    DF_ALL = lyriclensR:::read_all_lyrics(lyrics, write_output = TRUE)
+    DF_ALL = read_all_lyrics(lyrics, write_output = TRUE)
 
 
 
 # Search words ------------------------------------------------------------
 
   # Using json of a single Artist created by read_lyrics()
-  search_words(data = "outputs/lyrics/Lyrics_Tool.json",
+  search_words(lyrics = "outputs/lyrics/Lyrics_Tool.json",
                highlight_word = "love")
 
   # Using big DF with all lyrics created by read_all_lyrics()
   filename = here::here("outputs/DF_lyrics/DF_lyrics.gz")
   DF_ALL = data.table::fread(filename)
-  search_words(data = DF_ALL, highlight_word = "amor")
+  search_words(lyrics = DF_ALL, highlight_word = "amor")
 
 
 
@@ -87,22 +91,13 @@ devtools::load_all()
 
 # STEP BY STEP -------------------------------------------
 
-DF = read_lyrics(lyrics_file)
-DF_words = extract_words(DF)
+DF = read_lyrics(lyrics_file) # OR DF_ALL
+DF_words = extract_words(DF, ngram = 1)
 DF_freq = clean_words(DF_words)
 
+search_words(lyrics = lyrics_file, highlight_words = "love")
 
-
-DF_words = extract_words(DF_ALL, ngram = 1)
-DF_freq = clean_words(DF_words)
-
-
-
-search_words(data = DF_ALL,
-             highlight_words = "letra")
-
-
-DF_plot = DF_freq |> dplyr::filter(freq > 1000)
+DF_plot = DF_freq |> dplyr::filter(freq > 10) # If using DF_ALL increase freq
 
 set.seed(1)
 PLOT = wordcloud2::wordcloud2(DF_plot, size=1.6, color='random-dark')
@@ -113,7 +108,8 @@ PLOT
 # Search songs ------------------------------------------------------------
 
 # First steps towards looking for specific songs before downloading
-DF_songs_found = song_is_in_lyrics("amor")
+DF_songs_found = song_is_in_lyrics("amor",
+                                   filename = "outputs/DF_lyrics/DF_lyrics_ALL_es.gz")